move date progress to own component
